use long as the machine word width type
also, assume int is always 32-bit wide. This implies a lp64 ABI must be
used.

+++ src/arch/riscv/inc/arch/sbi.h
@@ -40,7 +40,7 @@ struct sbi_hsm {
     spinlock_t lock;
     enum { STARTED, STOPPED, START_PENDING, STOP_PENDING } state;
     vaddr_t start_addr;
-    uint64_t priv; 
+    unsigned priv;
 };
 
 void sbi_init();
