+++ src/arch/riscv/sbi.c
@@ -378,7 +378,7 @@ struct sbiret sbi_hsm_start_handler() {
                 ret.error = SBI_ERR_FAILURE;
             } else {
                 vaddr_t start_addr = vcpu_readreg(cpu.vcpu, REG_A1);
-                uint64_t priv = vcpu_readreg(cpu.vcpu, REG_A2);
+                unsigned priv = vcpu_readreg(cpu.vcpu, REG_A2);
                 vcpu->arch.sbi_ctx.state = START_PENDING;
                 vcpu->arch.sbi_ctx.start_addr = start_addr;
                 vcpu->arch.sbi_ctx.priv = priv;
@@ -440,9 +440,9 @@ struct sbiret sbi_bao_handler(unsigned long fid){
 
     struct sbiret ret;
 
-    uint64_t arg0 = vcpu_readreg(cpu.vcpu, REG_A0);
-    uint64_t arg1 = vcpu_readreg(cpu.vcpu, REG_A1);
-    uint64_t arg2 = vcpu_readreg(cpu.vcpu, REG_A2);
+    unsigned long arg0 = vcpu_readreg(cpu.vcpu, REG_A0);
+    unsigned long arg1 = vcpu_readreg(cpu.vcpu, REG_A1);
+    unsigned long arg2 = vcpu_readreg(cpu.vcpu, REG_A2);
 
     switch(fid) {
         case HC_IPC:
